requirements.txt
streamlit
pandas
camelot-py
opencv-python
openpyxl